+++ tests/playwright/sanity/includes/elements/section/reverse-columns/reverse-columns.test.ts
@@ -1,43 +1,72 @@
 import { test } from '@playwright/test';
 import Breakpoints from '../../../../../assets/breakpoints';
 import ReverseColumns from './reverse-columns';
+import WpAdminPage from '../../../../../pages/wp-admin-page';
 
 test.describe( 'Reverse Columns tests @reverse-columns', () => {
-	/**
-	 * Experiment Breakpoints(Off)
-	 */
-	for ( const testDevice of Breakpoints.getBasic() ) {
-		if ( 'desktop' === testDevice ) {
-			continue;
+	test.describe( 'Experiment Breakpoints: Off', () => {
+		test.beforeAll( async ( { browser }, testInfo ) => {
+			const context = await browser.newContext();
+
+			const page = await context.newPage();
+
+			const wpAdmin = new WpAdminPage( page, testInfo );
+			await wpAdmin.setExperiments( {
+				additional_custom_breakpoints: false,
+				container: false,
+			} );
+		} );
+		for ( const testDevice of Breakpoints.getBasic() ) {
+			if ( 'desktop' === testDevice ) {
+				continue;
+			}
+
+			test( `Reverse columns:${ testDevice } - Experiment breakpoints:Off`, async ( { page }, testInfo ) => {
+				const reverseColumns = new ReverseColumns( page, testInfo );
+				await reverseColumns.testReverseColumnsOneActivated( testDevice );
+			} );
 		}
 
-		test( `Reverse columns:${ testDevice } - Experiment breakpoints:Off`, async ( { page }, testInfo ) => {
+		test( 'Reverse columns:All - Experiment breakpoints:Off ', async ( { page }, testInfo ) => {
 			const reverseColumns = new ReverseColumns( page, testInfo );
-			await reverseColumns.testReverseColumnsOneActivated( testDevice );
+			await reverseColumns.testReverseColumnsAllActivated();
 		} );
-	}
-
-	test( 'Reverse columns:All - Experiment breakpoints:Off ', async ( { page }, testInfo ) => {
-		const reverseColumns = new ReverseColumns( page, testInfo );
-		await reverseColumns.testReverseColumnsAllActivated();
 	} );
 
-	/**
-	 * Experiment Breakpoints(On)
-	 */
-	for ( const testDevice of Breakpoints.getAll() ) {
-		if ( 'desktop' === testDevice ) {
-			continue;
+	test.describe( 'Experiment Breakpoints: On', () => {
+		test.beforeAll( async ( { browser }, testInfo ) => {
+			const context = await browser.newContext();
+			const page = await context.newPage();
+
+			const wpAdmin = new WpAdminPage( page, testInfo );
+			await wpAdmin.setExperiments( {
+				additional_custom_breakpoints: true,
+				container: false,
+			} );
+			const reverseColumns = new ReverseColumns( page, testInfo );
+			await reverseColumns.initAdditionalBreakpoints();
+		} );
+		test.afterAll( async ( { browser }, testInfo ) => {
+			const context = await browser.newContext();
+			const page = await context.newPage();
+
+			const reverseColumns = new ReverseColumns( page, testInfo );
+			await reverseColumns.resetAdditionalBreakpoints();
+		} );
+		for ( const testDevice of Breakpoints.getAll() ) {
+			if ( 'desktop' === testDevice ) {
+				continue;
+			}
+
+			test( `Reverse columns:${ testDevice } - Experiment breakpoints:On`, async ( { page }, testInfo ) => {
+				const reverseColumns = new ReverseColumns( page, testInfo );
+				await reverseColumns.testReverseColumnsOneActivated( testDevice, true );
+			} );
 		}
 
-		test( `Reverse columns:${ testDevice } - Experiment breakpoints:On`, async ( { page }, testInfo ) => {
+		test( 'Reverse columns:All - Experiment breakpoints:On', async ( { page }, testInfo ) => {
 			const reverseColumns = new ReverseColumns( page, testInfo );
-			await reverseColumns.testReverseColumnsOneActivated( testDevice, true );
+			await reverseColumns.testReverseColumnsAllActivated( true );
 		} );
-	}
-
-	test( 'Reverse columns:All - Experiment breakpoints:On', async ( { page }, testInfo ) => {
-		const reverseColumns = new ReverseColumns( page, testInfo );
-		await reverseColumns.testReverseColumnsAllActivated( true );
 	} );
 } );
