+++ tests/playwright/sanity/plugin-tester-container.test.ts-snapshots/editor-linux.png
