+++ tests/playwright/sanity/plugin-tester-section.test.ts
@@ -1,0 +1,16 @@
+import { test } from '@playwright/test';
+import { generatePluginTests } from './plugin-tester-generator';
+import WpAdminPage from '../pages/wp-admin-page';
+
+test.describe( `Plugin tester tests: sections`, () => {
+	test.beforeAll( async ( { browser }, testInfo ) => {
+		const page = await browser.newPage();
+		const wpAdmin = new WpAdminPage( page, testInfo );
+		await wpAdmin.resetExperiments();
+		await wpAdmin.setExperiments( { container: 'inactive' } );
+
+		await page.close();
+	} );
+
+	generatePluginTests( 'sections' );
+} );
